Move IndexRange.hpp to LayerRegion.hpp and fix warnings
